Add active styles to history period buttons
