+++ src/components/ExchangeRateHistory/ExchangeRateHistory.module.scss
@@ -1,0 +1,52 @@
+@import "../../styles/utils/variables";
+
+.historyRate {
+  text-align: center;
+  margin-bottom: 10px;
+}
+
+.periodButtons {
+  display: flex;
+  justify-content: center;
+}
+
+.monthButton, .monthButton:hover {
+  margin: 10px;
+  background-color: $cc-purple;
+  border-color: $cc-purple;
+
+  &.active {
+    background-color: $cc-purple-dark;
+  }
+}
+
+
+@media screen and (min-width: 414px) {
+  .backToDashboard {
+    margin-top: 15px
+  }
+}
+
+@media screen and (max-width: 414px) {
+  .monthButton {
+    height: 30px;
+    font-size: 0.8rem;
+  }
+}
+
+@media screen and (max-width: 365px) {
+  .monthButton {
+    margin: 10px 5px;
+
+    :global {
+      .anticon {
+        display: none
+      }
+    }
+
+    span:nth-child(2) {
+      margin: 0;
+    }
+  }
+}
+
