--- src/components/ExchangeRateHistoryComponent/ExchangeRateHistoryComponent.scss
@@ -1,62 +1,0 @@
-@import "../../styles/utils/variables";
-
-.history-rate-container {
-  text-align: center;
-  margin-bottom: 10px;
-
-  .periodButtons {
-    display: flex;
-    justify-content: center;
-
-    .monthButton {
-      margin: 10px;
-      background-color: $cc-purple;
-      border-color: $cc-purple;
-    }
-  }
-
-  .ant-spin {
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    height: 3rem;
-    width: 3rem;
-    margin: auto;
-  }
-
-  .backToDashboard {
-    margin-top: 15px
-  }
-}
-
-@media screen and (max-width: 414px) {
-  .periodButtons {
-    button {
-      height: 30px;
-      font-size: 0.8rem;
-    }
-  }
-
-  .backToDashboard {
-    margin-top: 0;
-  }
-}
-
-@media screen and (max-width: 365px) {
-  .periodButtons {
-    button {
-      margin-right: 0;
-
-      .anticon {
-        display: none
-      }
-
-      span:nth-child(2) {
-        margin: 0;
-      }
-    }
-  }
-}
-
